Refactor plant form with sensor and add sensor data display
The code refactors the plant form with sensor, incorporating a new cubit
to manage plant data, utilizing BlocProvider, and adding a new switch to
control plant activity. Additionally, it integrates the display of
sensor data using the SfCircularChart from the Syncfusion Flutter
package. This change sets the stage for improved management and
visualization of plant data with sensors. No associated issue.